+++ fractal/common/utils.py
@@ -1,0 +1,2 @@
+def slugify(value: str):
+    return value.lower().replace(" ", "_")
