+++ fractal/server/app/models/__init__.py
@@ -1,3 +1,4 @@
 from .project import *  # noqa: F403, F401
 from .security import *  # noqa: F403, F401
 from .task import *  # noqa: F403, F401
+from fractal.common.models import *  # noqa F401
